pkg/trace/pb: implement next version of encoder
