+++ pkg/trace/pb/decoder.go
@@ -7,12 +7,8 @@ package pb
 
 import (
 	"errors"
-	fmt "fmt"
-	io "io"
 	"math"
-	"sync"
 
-	"github.com/philhofer/fwd"
 	"github.com/tinylib/msgp/msgp"
 )
 
@@ -40,18 +36,6 @@ func parseString(dc *msgp.Reader) (string, error) {
 	default:
 		return "", msgp.TypeError{Encoded: t, Method: msgp.StrType}
 	}
-}
-
-// parseStringDict
-func parseStringDict(dc *msgp.Reader, dict []string) (string, error) {
-	i, err := dc.ReadInt()
-	if err != nil {
-		return "", err
-	}
-	if i > len(dict)-1 {
-		return "", fmt.Errorf("dictionary index %d out of place", i)
-	}
-	return dict[i], nil
 }
 
 // parseFloat64 parses a float64 even if the sent value is an int64 or an uint64;
@@ -214,215 +198,3 @@ func parseInt32(dc *msgp.Reader) (int32, error) {
 		return 0, msgp.TypeError{Encoded: t, Method: msgp.IntType}
 	}
 }
-
-// DecodeMsgArray implements msgp.Decodable
-func (z *Traces) DecodeMsgArray(dc *msgp.Reader) (err error) {
-	if _, err := dc.ReadArrayHeader(); err != nil {
-		return err
-	}
-	// read dictionary
-	sz, err := dc.ReadArrayHeader()
-	if err != nil {
-		return err
-	}
-	dict := make([]string, sz)
-	for i := range dict {
-		str, err := parseString(dc)
-		if err != nil {
-			return err
-		}
-		dict[i] = str
-	}
-	// read traces
-	var xsz uint32
-	xsz, err = dc.ReadArrayHeader()
-	if err != nil {
-		return
-	}
-	if cap((*z)) >= int(xsz) {
-		(*z) = (*z)[:xsz]
-	} else {
-		(*z) = make(Traces, xsz)
-	}
-	for wht := range *z {
-		var xsz uint32
-		xsz, err = dc.ReadArrayHeader()
-		if err != nil {
-			return
-		}
-		if cap((*z)[wht]) >= int(xsz) {
-			(*z)[wht] = (*z)[wht][:xsz]
-		} else {
-			(*z)[wht] = make(Trace, xsz)
-		}
-		for hct := range (*z)[wht] {
-			if (*z)[wht][hct] == nil {
-				(*z)[wht][hct] = new(Span)
-			}
-			err = (*z)[wht][hct].DecodeMsgArray(dc, dict)
-			if err != nil {
-				return
-			}
-		}
-	}
-	return
-}
-
-const spanPropertyCount = 12
-
-// DecodeMsgArray implements msgp.Decodable
-func (z *Span) DecodeMsgArray(dc *msgp.Reader, dict []string) (err error) {
-	var xsz uint32
-	xsz, err = dc.ReadArrayHeader()
-	if err != nil {
-		return
-	}
-	if xsz != spanPropertyCount {
-		return errors.New("encoded span needs exactly 12 elements in array")
-	}
-	for i := 0; i < spanPropertyCount; i++ {
-		if dc.IsNil() {
-			if err := dc.ReadNil(); err != nil {
-				return err
-			}
-			continue
-		}
-		switch i {
-		case 0:
-			// Service (0)
-			z.Service, err = parseStringDict(dc, dict)
-			if err != nil {
-				return
-			}
-		case 1:
-			// Name (1)
-			z.Name, err = parseStringDict(dc, dict)
-			if err != nil {
-				return
-			}
-		case 2:
-			// Resource (2)
-			z.Resource, err = parseStringDict(dc, dict)
-			if err != nil {
-				return
-			}
-		case 3:
-			// TraceID (3)
-			z.TraceID, err = parseUint64(dc)
-			if err != nil {
-				return
-			}
-		case 4:
-			// SpanID (4)
-			z.SpanID, err = parseUint64(dc)
-			if err != nil {
-				return
-			}
-		case 5:
-			// ParentID (5)
-			z.ParentID, err = parseUint64(dc)
-			if err != nil {
-				return
-			}
-		case 6:
-			// Start (6)
-			z.Start, err = parseInt64(dc)
-			if err != nil {
-				return
-			}
-		case 7:
-			// Duration (7)
-			z.Duration, err = parseInt64(dc)
-			if err != nil {
-				return
-			}
-		case 8:
-			// Error (8)
-			z.Error, err = parseInt32(dc)
-			if err != nil {
-				return
-			}
-		case 9:
-			// Meta (9)
-			var zwht uint32
-			zwht, err = dc.ReadMapHeader()
-			if err != nil {
-				return
-			}
-			if z.Meta == nil && zwht > 0 {
-				z.Meta = make(map[string]string, zwht)
-			} else if len(z.Meta) > 0 {
-				for key := range z.Meta {
-					delete(z.Meta, key)
-				}
-			}
-			for zwht > 0 {
-				zwht--
-				var zxvk string
-				var zbzg string
-				zxvk, err = parseStringDict(dc, dict)
-				if err != nil {
-					return
-				}
-				zbzg, err = parseStringDict(dc, dict)
-				if err != nil {
-					return
-				}
-				z.Meta[zxvk] = zbzg
-			}
-		case 10:
-			// Metrics (10)
-			var zhct uint32
-			zhct, err = dc.ReadMapHeader()
-			if err != nil {
-				return
-			}
-			if z.Metrics == nil && zhct > 0 {
-				z.Metrics = make(map[string]float64, zhct)
-			} else if len(z.Metrics) > 0 {
-				for key := range z.Metrics {
-					delete(z.Metrics, key)
-				}
-			}
-			for zhct > 0 {
-				zhct--
-				var zbai string
-				var zcmr float64
-				zbai, err = parseStringDict(dc, dict)
-				if err != nil {
-					return
-				}
-				zcmr, err = parseFloat64(dc)
-				if err != nil {
-					return
-				}
-				z.Metrics[zbai] = zcmr
-			}
-		case 11:
-			// Type (11)
-			z.Type, err = parseStringDict(dc, dict)
-			if err != nil {
-				return
-			}
-		}
-	}
-	return nil
-}
-
-var readerPool = sync.Pool{New: func() interface{} { return &msgp.Reader{} }}
-
-// NewMsgpReader returns a *msgp.Reader that
-// reads from the provided reader. The
-// reader will be buffered.
-func NewMsgpReader(r io.Reader) *msgp.Reader {
-	p := readerPool.Get().(*msgp.Reader)
-	if p.R == nil {
-		p.R = fwd.NewReader(r)
-	} else {
-		p.R.Reset(r)
-	}
-	return p
-}
-
-// FreeMsgpReader marks reader r as done.
-func FreeMsgpReader(r *msgp.Reader) { readerPool.Put(r) }
